initial commit in coffee shop

+++ src/components/Users.jsx
@@ -1,56 +1,73 @@
 import { useState } from "react";
 import { useLoaderData } from "react-router-dom";
-
+import Swal from "sweetalert2";
 
 const Users = () => {
-    const loadedUsers = useLoaderData();
-    const [users, setUsers] = useState(loadedUsers);
+    const loadedUser = useLoaderData();
+    const [users, setUsers] = useState(loadedUser);
 
-    const handleDelete = id => {
-        // make sure user is confirmed to delete
-        fetch(`https://coffee-store-server-74xiae2di-jhankarphero.vercel.app/user/${id}`, {
-            method: 'DELETE'
+    const handleDeleteUser = id => {
+
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+        }).then((result) => {
+            if (result.isConfirmed) {
+
+                fetch(`http://localhost:5000/user/${id}`, {
+                    method: 'DELETE'
+                })
+                    .then(res => res.json())
+                    .then(data => {
+                        Swal.fire(
+                            'Deleted!',
+                            'Your file has been deleted.',
+                            'success'
+                        )
+                        console.log(data)
+                        if (data.deletedCount > 0) {
+                            const remaining = users?.filter(user => user._id !== id)
+                            setUsers(remaining)
+                        }
+                    })
+
+                // 
+            }
         })
-            .then(res => res.json())
-            .then(data => {
-                if (data.deletedCount > 0) {
-                    console.log('deleted successfully');
-                    // remove the user from the UI
-                    const remainingUsers = users.filter(user => user._id !== id);
-                    setUsers(remainingUsers);
-                }
-            })
-    }
 
+
+
+    }
     return (
         <div>
-            <h2>Users: {loadedUsers.length}</h2>
+            <h2>Users: {loadedUser.length}</h2>
             <div className="overflow-x-auto">
                 <table className="table">
                     {/* head */}
                     <thead>
                         <tr>
                             <th></th>
-                            <th>Email</th>
-                            <th>Created At</th>
-                            <th>Last Logged In</th>
+                            <th>Name</th>
+                            <th>Job</th>
+                            <th>LoggedIn</th>
                             <th>Action</th>
                         </tr>
                     </thead>
                     <tbody>
                         {
-                            users.map(user => <tr key={user._id}>
-                                <th>1</th>
+                            users?.map((user, i) => <tr key={user._id} className="bg-base-200">
+                                <th>{i + 1}</th>
                                 <td>{user.email}</td>
                                 <td>{user.createdAt}</td>
                                 <td>{user.lastLoggedAt}</td>
-                                <td>
-                                    <button
-                                        onClick={() => handleDelete(user._id)} className="btn">X</button>
-                                </td>
+                                <td><button className="btn btn-warning" onClick={() => handleDeleteUser(user._id)}>X</button></td>
                             </tr>)
                         }
-
                     </tbody>
                 </table>
             </div>
